Refactor MonnigAdmin to Curator
