+++ src/main/java/edu/tcu/cs/monnigmeteoritecollection/curator/CuratorRepository.java
@@ -1,5 +1,5 @@
-package edu.tcu.cs.monnigmeteoritecollection.monnigadmin;
+package edu.tcu.cs.monnigmeteoritecollection.curator;
 
-public class AdminRepository {
+public class CuratorRepository {
     
 }
